backend/requirements.txt
flask
flask_cors
python-dotenv
pandas
numpy
matplotlib
sklearn